split tools.ts into individual files
